backend/requirements.txt
fastapi
python-dotenv
uvicorn
websockets
openai
requests